requirements.txt
qdrant-client
pypdf
transformers
torch
llama-index-llms-ollama
llama-index
llama-index-vector-stores-qdrant
